split apart api_handlers_streaming
